Query/Command Repository 분리

+++ src/main/java/com/buckwheat/garden/data/dto/chemical/ChemicalDetail.java
@@ -8,5 +8,5 @@ import lombok.*;
 @AllArgsConstructor
 public class ChemicalDetail {
     private ChemicalDto chemical;
-    private int wateringSize;
+    private Long wateringSize;
 }
